pixels are getting pulled all the way through now, but ycbcr 10bit is... off

+++ CapturePlugins/TanagerPlugin/ComputeShaders/Linearize_ITUR_BT709.hlsl
@@ -1,5 +1,5 @@
-RWTexture2D<uint4> inputTexture : register(u0); // 16bpc uint 444 texture
-RWTexture2D<uint4> outputTexture : register(u1); // 16bpc uint 444 texture
+RWTexture2D<float4> inputTexture : register(u0); // 16bpc uint 444 texture
+RWTexture2D<float4> outputTexture : register(u1); // 16bpc uint 444 texture
 
 [numthreads(1, 1, 1)]
 void main(uint3 DTid : SV_DispatchThreadID)
